LayoutTests/imported/w3c/web-platform-tests/editing/other/insertparagraph-in-inline-editing-host.tentative_white-space=pre&display=inline-block&method=shift-enter-expected.txt

abcd￡

FAIL insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">{}</span> (defaultParagraphSeparator=div) assert_in_array: A linefeed and additional line breaker should be inserted when insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">{}</span> (defaultParagraphSeparator=div) value "<span contenteditable=\"\">￡\n\n</span>" not in array ["<span contenteditable=\"\">\n<br></span>", "<span contenteditable=\"\">\n\n</span>"]
FAIL insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">{}</span> followed by a <br> (defaultParagraphSeparator=div) assert_in_array: A linefeed and additional line break should be inserted when insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">{}</span> followed by a <br> (defaultParagraphSeparator=div) value "<span contenteditable=\"\">￡\n\n</span><br>" not in array ["<span contenteditable=\"\">\n\n</span><br>", "<span contenteditable=\"\">\n<br></span><br>"]
FAIL insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">{}</span> followed by text (defaultParagraphSeparator=div) assert_in_array: A linefeed and additional line break should be inserted when insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">{}</span> followed by text (defaultParagraphSeparator=div) value "<span contenteditable=\"\">￡\n\n</span>abc" not in array ["<span contenteditable=\"\">\n\n</span>abc", "<span contenteditable=\"\">\n<br></span>abc"]
FAIL insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">{}<br></span> (defaultParagraphSeparator=div) assert_equals: A <br> should be inserted when insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">{}<br></span> (defaultParagraphSeparator=div) expected "<span contenteditable=\"\">\n<br></span>" but got "<span contenteditable=\"\">￡\n<br></span>"
FAIL insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">[]abcd</span> (defaultParagraphSeparator=div) assert_equals: A linefeed should be inserted when insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">[]abcd</span> (defaultParagraphSeparator=div) expected "<span contenteditable=\"\">\nabcd</span>" but got "<span contenteditable=\"\">￡\nabcd</span>"
FAIL insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">ab[]cd</span> (defaultParagraphSeparator=div) assert_equals: A linefeed should be inserted when insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">ab[]cd</span> (defaultParagraphSeparator=div) expected "<span contenteditable=\"\">ab\ncd</span>" but got "<span contenteditable=\"\">ab￡\ncd</span>"
FAIL insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">abcd[]</span> (defaultParagraphSeparator=div) assert_in_array: A linefeed and additional line break should be inserted when insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">abcd[]</span> (defaultParagraphSeparator=div) value "<span contenteditable=\"\">abcd￡\n\n</span>" not in array ["<span contenteditable=\"\">abcd\n<br></span>", "<span contenteditable=\"\">abcd\n\n</span>"]
FAIL insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">abcd[]</span> followed by a <br> element (defaultParagraphSeparator=div) assert_in_array: A linefeed and additional line break should be inserted when insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">abcd[]</span> followed by a <br> element (defaultParagraphSeparator=div) value "<span contenteditable=\"\">abcd￡\n\n</span><br>" not in array ["<span contenteditable=\"\">abcd\n<br></span><br>", "<span contenteditable=\"\">abcd\n\n</span><br>"]
FAIL insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">abcd[]</span> followed by text (defaultParagraphSeparator=div) assert_in_array: A linefeed and additional line break should be inserted when insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">abcd[]</span> followed by text (defaultParagraphSeparator=div) value "<span contenteditable=\"\">abcd￡\n\n</span>efg" not in array ["<span contenteditable=\"\">abcd\n<br></span>efg", "<span contenteditable=\"\">abcd\n\n</span>efg"]
FAIL insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">{}</span> (defaultParagraphSeparator=p) assert_in_array: A linefeed and additional line breaker should be inserted when insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">{}</span> (defaultParagraphSeparator=p) value "<span contenteditable=\"\">￡\n\n</span>" not in array ["<span contenteditable=\"\">\n<br></span>", "<span contenteditable=\"\">\n\n</span>"]
FAIL insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">{}</span> followed by a <br> (defaultParagraphSeparator=p) assert_in_array: A linefeed and additional line break should be inserted when insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">{}</span> followed by a <br> (defaultParagraphSeparator=p) value "<span contenteditable=\"\">￡\n\n</span><br>" not in array ["<span contenteditable=\"\">\n\n</span><br>", "<span contenteditable=\"\">\n<br></span><br>"]
FAIL insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">{}</span> followed by text (defaultParagraphSeparator=p) assert_in_array: A linefeed and additional line break should be inserted when insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">{}</span> followed by text (defaultParagraphSeparator=p) value "<span contenteditable=\"\">￡\n\n</span>abc" not in array ["<span contenteditable=\"\">\n\n</span>abc", "<span contenteditable=\"\">\n<br></span>abc"]
FAIL insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">{}<br></span> (defaultParagraphSeparator=p) assert_equals: A <br> should be inserted when insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">{}<br></span> (defaultParagraphSeparator=p) expected "<span contenteditable=\"\">\n<br></span>" but got "<span contenteditable=\"\">￡\n<br></span>"
FAIL insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">[]abcd</span> (defaultParagraphSeparator=p) assert_equals: A linefeed should be inserted when insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">[]abcd</span> (defaultParagraphSeparator=p) expected "<span contenteditable=\"\">\nabcd</span>" but got "<span contenteditable=\"\">￡\nabcd</span>"
FAIL insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">ab[]cd</span> (defaultParagraphSeparator=p) assert_equals: A linefeed should be inserted when insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">ab[]cd</span> (defaultParagraphSeparator=p) expected "<span contenteditable=\"\">ab\ncd</span>" but got "<span contenteditable=\"\">ab￡\ncd</span>"
FAIL insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">abcd[]</span> (defaultParagraphSeparator=p) assert_in_array: A linefeed and additional line break should be inserted when insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">abcd[]</span> (defaultParagraphSeparator=p) value "<span contenteditable=\"\">abcd￡\n\n</span>" not in array ["<span contenteditable=\"\">abcd\n<br></span>", "<span contenteditable=\"\">abcd\n\n</span>"]
FAIL insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">abcd[]</span> followed by a <br> element (defaultParagraphSeparator=p) assert_in_array: A linefeed and additional line break should be inserted when insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">abcd[]</span> followed by a <br> element (defaultParagraphSeparator=p) value "<span contenteditable=\"\">abcd￡\n\n</span><br>" not in array ["<span contenteditable=\"\">abcd\n<br></span><br>", "<span contenteditable=\"\">abcd\n\n</span><br>"]
FAIL insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">abcd[]</span> followed by text (defaultParagraphSeparator=p) assert_in_array: A linefeed and additional line break should be inserted when insertLineBreak in <span contenteditable style="display:inline-block;white-space:pre">abcd[]</span> followed by text (defaultParagraphSeparator=p) value "<span contenteditable=\"\">abcd￡\n\n</span>efg" not in array ["<span contenteditable=\"\">abcd\n<br></span>efg", "<span contenteditable=\"\">abcd\n\n</span>efg"]
